pyproject.toml
[tool.poetry]
name = "ag-smile-leaseback-crm-back"
version = "0.1.0"
description = ""
authors = ["Aiful Inc, <engineering@pj100.biz>"]
readme = "README.md"
packages = [{include = "ag_smile_leaseback_crm_back"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
Django = "^4.1.2"
mysqlclient = "^2.1.1"
gunicorn = "^20.1.0"
djangorestframework = "^3.14.0"
drf-nested-routers = "^0.93.4"
drf-spectacular = "^0.24.2"
pytest = "^7.1.3"
django-cors-headers = "^3.13.0"
Authlib = "^1.1.0"
django-filter = "^22.1"
django-ses = "^3.2.2"
django-storages = "^1.13.1"
django-debug-toolbar = "^3.7.0"
django-extensions = "^3.2.1"
pillow="^9.3.0"
pytest-cov = "^4.0.0"
pytest-sugar = "^0.9.6"
pytest-django = "^4.5.2"
allure-pytest = "^2.12.0"
boto3 = "^1.26.17"
pytest-custom-exit-code = "^0.3.0"

[tool.poetry.group.dev.dependencies]
black = {version = "^22.10.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
